IMP - Refactor ATK skill sorter file structure
