+++ src/components/elements/gameData/skillAtk/sorter/main.tsx
@@ -3,8 +3,8 @@ import React from 'react';
 import Dropdown from 'react-bootstrap/Dropdown';
 import DropdownButton from 'react-bootstrap/DropdownButton';
 
-import {useI18n} from '../../../../../../i18n/hook';
-import {InputData, SortBy} from '../../in/types';
+import {useI18n} from '../../../../../i18n/hook';
+import {InputData, SortBy} from '../in/types';
 import {orderName} from './lookup';
